Move final-evals secrets manager to pipeline/common
This allows reusing it in dataset tasks for HF auth
