+++ pipeline/data/parallel_downloaders.py
@@ -39,6 +39,14 @@ def huggingface(src: LangCode, trg: LangCode, dataset: str, output_prefix: Path)
     parsed = HFDATASET_PARSE.match(dataset)
     if not parsed:
         raise ValueError(f"Could not parse HF dataset '{dataset}'")
+
+    # Log in to Huggingface for gated datasets and rate limiting
+    if os.environ.get("TASK_ID"):
+        from pipeline.common.secrets import Secrets
+
+        secrets = Secrets()
+        secrets.prepare_key_hf()
+
     # import inline because otherwise datasets needs to be added to pyproject
     # and it's difficult to lock
     from datasets import load_dataset  # pyright: ignore [reportMissingImports]
