+++ pipeline/data/requirements/data.in
@@ -7,3 +7,4 @@ OpenCC==1.1.9
 sacrebleu==2.4.2
 PyICU==2.15.2
 datasets==4.8.4
+taskcluster==88.0.2
