+++ pipeline/eval/final_eval.py
@@ -18,7 +18,6 @@ import gc
 import json
 import logging
 import os
-import tempfile
 import time
 from collections import defaultdict
 from dataclasses import dataclass
@@ -387,32 +386,6 @@ class Storage:
         return str(name).replace("/", self.SEPARATOR)
 
 
-class Secrets:
-    def __init__(self):
-        import taskcluster
-
-        root_url = os.environ.get("TASKCLUSTER_PROXY_URL")
-        assert root_url, "When running in Taskcluster the TASKCLUSTER_PROXY_URL must be set."
-        self.secrets = taskcluster.Secrets({"rootUrl": root_url})
-
-    def prepare_keys(self):
-        logger.info("Reading secrets from Taskcluster")
-        os.environ["HF_TOKEN"] = self.read_key("huggingface")["token"]
-        os.environ["OPENAI_API_KEY"] = self.read_key("chatgpt")["token"]
-        os.environ["AZURE_TRANSLATOR_KEY"] = self.read_key("azure-translate")["token"]
-        google_key_file = tempfile.NamedTemporaryFile("w", delete=False)
-        with google_key_file:
-            json.dump(self.read_key("google-translate"), google_key_file)
-        os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = google_key_file.name
-
-    def read_key(self, name: str) -> dict:
-        try:
-            response = self.secrets.get(f"project/translations/level-1/{name}")
-            return response["secret"]
-        except Exception as e:
-            raise ValueError(f"Could not retrieve the secret key {name}: {e}")
-
-
 class EvalsRunner:
     def __init__(self, config: Config):
         logger.info(f"Config: {config.print()}")
@@ -424,6 +397,8 @@ class EvalsRunner:
         self.run_timestamp = datetime.datetime.utcnow().strftime("%Y%m%dT%H%M%S")
         # Check if we're inside a Taskcluster task
         if os.environ.get("TASK_ID"):
+            from pipeline.common.secrets import Secrets
+
             secrets = Secrets()
             secrets.prepare_keys()
